Move tests to the v3 package
